refactor: 사용자 통계 기능 리팩토링  - DailyRecordResponse 데이터 타입 수정  - UserStatController의getAverageFocusRatio 반환 데이터 수정  - AverageFocusRatioResponse 생성

+++ livestudy/src/main/java/org/livestudy/dto/DailyRecordResponse.java
@@ -9,7 +9,7 @@ import java.time.LocalDate;
 @Builder
 public class DailyRecordResponse {
     private LocalDate recordDate;
-    private int dailyStudyTime;
-    private int dailyAwayTime;
-    private double focusRatio;
+    private Long dailyStudyTime;
+    private Long dailyAwayTime;
+    private Double focusRatio;
 }
